+++ src/main/java/com/example/storyteller/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.example.storyteller.repository;
 import com.example.storyteller.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
 }
